refactor: [#190] AlertScene 분리 및 프로토콜 띄어쓰기 변경
