+++ FullStackCodingBot/FullStackCodingBot/View/Scene/SceneType/SceneType.swift
@@ -1,6 +1,5 @@
 import UIKit
 
 protocol SceneType {
-    
     func instantiate(from identifier: String) -> UIViewController
 }
